product_recognition.py
import os
import logging
import numpy as np
# Removing TensorFlow import to simplify the application
# import tensorflow as tf
# from tensorflow import keras
from PIL import Image
import io
import random

logger = logging.getLogger(__name__)

# Global variables
model = None
class_names = [
    "Coca-Cola", "Pepsi", "Dr Pepper", "Mountain Dew", "Doritos", 
    "Lay's", "Cheetos", "Oreo", "Hershey's", "Reese's", 
    "Campbell's Soup", "Heinz Ketchup", "Kraft Mac & Cheese", "Cheerios", "Lucky Charms"
]

def load_model():
    """
    Load or create the product recognition model.
    
    In a real production app, this would load a pre-trained model.
    For this example, we create a simple mock model to simulate recognition.
    """
    global model
    
    try:
        # For this example, we'll just use a mock model
        # In a real application, you would load a pre-trained TensorFlow model
        model = True  # Just a placeholder to indicate model is loaded
        
        logger.info("Product recognition model initialized successfully")
    except Exception as e:
        logger.error(f"Error loading model: {str(e)}")
        model = None

def preprocess_image(image):
    """
    Preprocess the image for the model.
    
    Args:
        image: PIL Image
        
    Returns:
        Preprocessed image as numpy array
    """
    try:
        # Resize the image to the required input dimensions
        image = image.resize((224, 224))
        
        # Convert to numpy array
        img_array = np.array(image) / 255.0
        
        # Expand dimensions to create batch
        img_array = np.expand_dims(img_array, axis=0)
        
        return img_array
    except Exception as e:
        logger.error(f"Error preprocessing image: {str(e)}")
        return None

def classify_product(image):
    """
    Classify the product in the image.
    
    Args:
        image: PIL Image
        
    Returns:
        Product name or None if no product is recognized
    """
    global model, class_names
    
    try:
        if model is None:
            logger.error("Model not loaded")
            return None
            
        # Preprocess the image
        processed_image = preprocess_image(image)
        
        if processed_image is None:
            return None
            
        # For demonstration purposes, we'll return a random product
        # In a real application, you'd use model.predict(processed_image)
        # and return the class with the highest probability
        
        # Simulate model prediction
        # In a real app, you would use:
        # predictions = model.predict(processed_image)
        # predicted_class_index = np.argmax(predictions[0])
        # But for this demo, we'll just randomly select a product
        predicted_class_index = np.random.randint(0, len(class_names))
        
        return class_names[predicted_class_index]
    except Exception as e:
        logger.error(f"Error classifying image: {str(e)}")
        return None
